Removing warning edit
I forked your fork because I cannot figure out how to change this otherwise. The common use cases are included below and the longer edit makes this harder to read in its intended use (held up alongside while the trainer is going through the content). Really working out the query is beyond the scope of 101, but will 100% make a good foundation for an Insights-based training at a future point.